WIP - extract structs from GatewayDNSReconciler
