+++ gameplan
@@ -6,9 +6,17 @@ x - Deploy contour in make e2e-up
 x - Make a deployment manifest for capi-dns-controller
 x   - rbac
 x - Deploy capi-dns-controller in make e2e-up
-  - Use metallb in our e2e tests
+x - Use metallb in our e2e tests
     - strict ARP mode?
-  - Implement controller
+> - Implement controller
+    - extract objects/functions to make smaller parts and improve test coverage
+       - perhaps each private method the Reconcile calls is public method on one or more new objects?
+    - should "writeEndpointSlicesToCluster" write to "nonMatching" clusters too?
+    - reconcile needs to remove endpoint slices (that we manage) when things (like what?) get deleted
+      - gatewaydns
+      - contour/envoy/lb ip
+    - reconcile needs to update endpoint slices (that we manage) when things change like:
+      - lb ip address changes
   - create "make" task to run counterfeiter (or any go generate task)
     - consider reusing make generate for this
   - mgmt-cluster namespace tenancy:
@@ -43,17 +51,3 @@ spec:
 
 - watch service
   - maybe this can trigger the gatewaydnsrecord reconcile somehow?
-
-
-
-Mon afternoon observations:
-- When a cluster has no services *with external ips*, it should not create an
-  enpoint slice for distribution to the workload clusters.  Right now the
-  controller attempts to create an endpoint slice with a null address. The e2e
-  test is failing because the controller is failing to write these bogus endpoint slices
-  to workload clusters. I failed to write a failing controller test to expose this
-  problem. The test at the bottom of
-  pkg/controllers/gatewaydns/controller_test.go should be failing but is not.
-  The controller has a lot of functionality, and probably should be broken up
-  so it's easier to test.
-
